Call controllers in parallel rather than nesting them
Originally, when a controller registered itself, it was inserted in the 
app as middleware. This not only led to a bunch of unnecessary levels of 
nesting, it also led to strange unintended interactions between 
controllers (such as one controller's middleware affecting the others'). 
It also made error handling quite a mess.

This changes controllers to be run in parallel, similar to 
Rack::Builder's `map` method. This allows for a much cleaner and 
easier-to-understand call tree (e.g. the app is always the caller for 
the controller) as well as allowing each controller to then have its own 
independent middleware chain. If 'global middleware is still needed, it 
can still be added on either the base controller or base app.

There don't seem to be any deep ramifications to this approach... 
everything seems to function as intended at the Sinatra level.