Instantiate controllers at request time
This is undoubtedly less performant than the way Sinatra normally does 
it, but there doesn't appear to be a way to maintain the middleware 
stack *and* also notify the controller of the current request's app 
instance. Brief testing didn't show that big a difference in 
performance, and given that the applications are generally quite 
lightweight, I don't foresee many problems as a result of this approach.